refactor: move standalone KV value types into submodule
KvEntry, VersionStamp and KvU64 are now in denokv._kv_values. These
types are needed by other submodules that I'd like to separate from the
denokv.kv module.
