+++ test/denokv_testing.py
@@ -29,6 +29,9 @@ from denokv._datapath_pb2 import MutationType
 from denokv._datapath_pb2 import ReadRange
 from denokv._datapath_pb2 import ReadRangeOutput
 from denokv._datapath_pb2 import ValueEncoding
+from denokv._kv_values import KvEntry
+from denokv._kv_values import KvU64
+from denokv._kv_values import VersionStamp
 from denokv._pycompat.dataclasses import slots_if310
 from denokv._pycompat.protobuf import enum_name
 from denokv._pycompat.typing import Any
@@ -49,11 +52,8 @@ from denokv.datapath import pack_key
 from denokv.datapath import parse_protobuf_kv_entry
 from denokv.errors import InvalidCursor
 from denokv.kv import AnyCursorFormat
-from denokv.kv import KvEntry
-from denokv.kv import KvU64
 from denokv.kv import LimitExceededPolicy
 from denokv.kv import ListContext
-from denokv.kv import VersionStamp
 from denokv.kv_keys import KvKey
 from denokv.result import Err
 from denokv.result import Ok
